+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/servlets/action/ImportDictionaryServlet.java
@@ -23,6 +23,8 @@ import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.text.MessageFormat;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -37,7 +39,6 @@ import static be.orbinson.aem.dictionarytranslator.utils.DictionaryConstants.SLI
 import static com.day.cq.commons.jcr.JcrConstants.JCR_LANGUAGE;
 import static com.day.cq.commons.jcr.JcrConstants.JCR_MIXINTYPES;
 import static com.day.cq.commons.jcr.JcrConstants.JCR_PRIMARYTYPE;
-import static com.day.cq.commons.jcr.JcrConstants.NT_UNSTRUCTURED;
 import static org.apache.sling.jcr.resource.api.JcrResourceConstants.NT_SLING_FOLDER;
 import static org.apache.sling.jcr.resource.api.JcrResourceConstants.SLING_RESOURCE_TYPE_PROPERTY;
 
@@ -85,7 +86,9 @@ public class ImportDictionaryServlet extends SlingAllMethodsServlet {
             validateCsvHeaders(response, headers);
 
             headers.remove(KEY_HEADER);
-            initializeLanguageData(headers, languages, translations);
+            Resource dictionary = resourceResolver.getResource(path);
+            Iterator<Resource> knownLanguages = dictionary.listChildren();
+            initializeLanguageData(headers, languages, translations, knownLanguages, response);
 
             for (CSVRecord record : csvParser) {
                 processCsvRecord(path, languages, resourceResolver, keys, translations, record);
@@ -118,10 +121,19 @@ public class ImportDictionaryServlet extends SlingAllMethodsServlet {
         }
     }
 
-    private void initializeLanguageData(Map<String, Integer> headers, List<String> languages, List<List<String>> translations) {
+    private void initializeLanguageData(Map<String, Integer> headers, List<String> languages, List<List<String>> translations, Iterator<Resource> knownLanguages, SlingHttpServletResponse response) {
         for (String language : headers.keySet()) {
-            languages.add(language);
-            translations.add(new ArrayList<>());
+            while (knownLanguages.hasNext()){
+                if (knownLanguages.next().getName().equals(language)){
+                    languages.add(language);
+                    translations.add(new ArrayList<>());
+                    break;
+                } else {
+                    String error = "Incorrect CSV file, please only add languages that exist in the dictionary";
+                    LOG.warn(error);
+                    response.setStatus(400, error);
+                }
+            }
         }
     }
 
@@ -148,11 +160,12 @@ public class ImportDictionaryServlet extends SlingAllMethodsServlet {
         if (languageResource == null) {
             languageResource = createLanguageResource(resourceResolver, path, language);
         }
-
-        String newNodeName = label;
-        Resource labelResource = createLabelResource(resourceResolver, languageResource, newNodeName);
-
-        updateLabelResourceProperties(labelResource, label, translation);
+        Resource labelResource = resourceResolver.getResource(path + "/" + language + "/" + label);
+        if (labelResource == null) {
+            createLabelResource(resourceResolver, languageResource, label, translation);
+        } else {
+            updateLabelResourceProperties(labelResource, label, translation);
+        }
     }
 
     private Resource createLanguageResource(ResourceResolver resourceResolver, String path, String language) throws PersistenceException {
@@ -165,8 +178,14 @@ public class ImportDictionaryServlet extends SlingAllMethodsServlet {
         ));
     }
 
-    private Resource createLabelResource(ResourceResolver resourceResolver, Resource languageResource, String newNodeName) throws PersistenceException {
-        return resourceResolver.create(languageResource, newNodeName, Map.of(JCR_PRIMARYTYPE, NT_UNSTRUCTURED));
+    private Resource createLabelResource(ResourceResolver resourceResolver, Resource languageResource, String newNodeName, String translation) throws PersistenceException {
+        Map<String, Object> properties = new HashMap<>();
+        properties.put(JCR_PRIMARYTYPE, SLING_MESSAGEENTRY);
+        properties.put(SLING_KEY, newNodeName);
+        if (!translation.isBlank()) {
+            properties.put(SLING_MESSAGE, translation);
+        }
+        return resourceResolver.create(languageResource, newNodeName, properties);
     }
 
     private void updateLabelResourceProperties(Resource labelResource, String label, String translation) {