+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+### Added
+
+- CSV files with other languages than the dictionary currently has will not be imported.
+
 ## [1.1.0] - 2024-02-01
 
 ### Changed